day2/requirements.txt
fastapi
uvicorn
python-dotenv
murf
